# dootask-plugin/openapi.yaml
openapi: 3.0.3
info:
  title: KPI API
  version: 1.0.0
  description: |
    DooTask KPI 绩效考核插件后端 API。

    说明：
    - 所有路径以 `/api/` 开头（Gin 在 `/api` 下注册，nginx 将 `/apps/kpi/api` 反代到 `kpi:8080/api`）。
    - 鉴权请求头（Token、X-Doo-User-Id、Authorization 等）由 doo app call 自动注入，无需在指令中传递，故规范中不声明这些 header。
    - 通过 `doo app call kpi <operationId>` 调用；query/path/扁平请求体字段会成为指令参数。
    - 内部回调（系统 hook、登录、SSE 流、文件下载、备份/恢复等）已用 `x-doo-cli: false` 隐藏。
...
paths:
...
  /api/employees:
    get:
      operationId: getEmployees
      summary: 获取员工列表
      tags: [employees]
      parameters:
        - { name: page, in: query, schema: { type: integer, default: 1 } }
        - { name: pageSize, in: query, schema: { type: integer, default: 10 } }
        - { name: search, in: query, schema: { type: string } }
        - { name: department_id, in: query, schema: { type: integer } }
        - { name: role, in: query, schema: { type: string, enum: [employee, manager, hr, "manager,hr"] } }
        - { name: is_active, in: query, schema: { type: boolean } }
      responses:
        '200':
          description: 员工分页列表
    post:
      operationId: createEmployee
      summary: 创建员工（HR/管理员）
      tags: [employees]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EmployeeInput'
      responses:
        '201':
          description: 创建成功
  /api/employees/{id}:
    get:
      operationId: getEmployee
      summary: 获取单个员工
      tags: [employees]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        '200':
          description: 员工详情
    put:
      operationId: updateEmployee
      summary: 更新员工（HR/管理员）
      tags: [employees]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EmployeeInput'
      responses:
        '200':
          description: 更新成功
    delete:
      operationId: deleteEmployee
      summary: 删除员工（HR）
      tags: [employees]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        '200':
          description: 删除成功
  /api/employees/{id}/subordinates:
    get:
      operationId: getEmployeeSubordinates
      summary: 获取员工的下属列表
      tags: [employees]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        '200':
          description: 下属列表
...
components:
  schemas:
...
    EmployeeInput:
      type: object
      properties:
        name: { type: string }
        email: { type: string, format: email }
        password: { type: string, minLength: 6, description: 仅编辑员工时填写，留空表示不修改 }
        position: { type: string }
        department_id: { type: integer }
        manager_id: { type: integer, nullable: true }
        role: { type: string, enum: [employee, manager, hr] }
        is_active: { type: boolean }
...